+++ gpt_langchain.py
